requirements.txt
Click==7.0
Flask==1.0.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.2
MarkupSafe==1.1.1
Pillow==5.4.1
pip==9.0.1
python-docx==0.8.10
scikit-surprise==1.0.6
scipy==1.1.0
setuptools==28.8.0
surprise==0.1
Werkzeug==0.14.1
WTForms==2.2.1

www.fae.cn/ht/downloadDetail.do?id=
